text overflow processing fixed

+++ aui.views/src/AUI/View/AAbstractLabel.h
@@ -47,7 +47,7 @@ private:
     void processTextOverflow(AString &text);
 
     template < class Iterator >
-    int findFirstOverflowedIndex(const Iterator& begin, const Iterator& end, int overflowingWidth);
+    size_t findFirstOverflowedIndex(const Iterator& begin, const Iterator& end, int overflowingWidth);
 
     template < class Iterator >
     void processTextOverflow(Iterator begin, Iterator end, int overflowingWidth);
